refactor egs/upit
