+++ egs/upit/README.md
@@ -1,3 +1,3 @@
-A simple uPIT-(B)LSTM egs to illustate usage of some scripts under [scripts](../../scripts)
+A simple uPIT-(B)LSTM example to illustate usage of some scripts under [scripts](../../scripts)
 
 See details in local/*.sh
