+++ egs/upit/nnet/libs/trainer.py
@@ -14,7 +14,7 @@ import torch.nn.functional as F
 from torch.optim.lr_scheduler import ReduceLROnPlateau
 from torch.nn.utils import clip_grad_norm_
 
-from utils import get_logger
+from .utils import get_logger
 
 
 def load_obj(obj, device):
@@ -33,35 +33,6 @@ def load_obj(obj, device):
         return cuda(obj)
 
 
-class AverageMeter(object):
-    """
-    A simple average meter
-    """
-
-    def __init__(self):
-        self.val = defaultdict(float)
-        self.cnt = defaultdict(int)
-
-    def reset(self):
-        self.val.clear()
-        self.cnt.clear()
-
-    def add(self, key, value):
-        self.val[key] += value
-        self.cnt[key] += 1
-
-    def value(self, key):
-        if self.cnt[key] == 0:
-            return 0
-        return self.val[key] / self.cnt[key]
-
-    def sum(self, key):
-        return self.val[key]
-
-    def count(self, key):
-        return self.cnt[key]
-
-
 class SimpleTimer(object):
     """
     A simple timer
@@ -77,26 +48,57 @@ class SimpleTimer(object):
         return time.time() - self.start
 
 
+class ProgressReporter(object):
+    """
+    A simple progress reporter
+    """
+
+    def __init__(self, logger, period=100):
+        self.period = period
+        self.logger = logger
+        self.loss = []
+        self.timer = SimpleTimer()
+
+    def add(self, loss):
+        self.loss.append(loss)
+        N = len(self.loss)
+        if not N % self.period:
+            avg = sum(self.loss[-self.period:]) / self.period
+            self.logger.info("Processed {:d} batches"
+                             "(loss = {:.2f})...".format(N, avg))
+
+    def report(self, details=False):
+        N = len(self.loss)
+        if details:
+            sstr = ",".join(map(lambda f: "{:.2f}".format(f), self.loss))
+            self.logger.info("Loss on {:d} batches: {}".format(N, sstr))
+        return {
+            "loss": sum(self.loss) / N,
+            "batches": N,
+            "cost": self.timer.elapsed()
+        }
+
+
 class Trainer(object):
     def __init__(self,
                  nnet,
                  checkpoint="checkpoint",
                  optimizer="adam",
-                 gpuid="0",
+                 gpuid=0,
                  optimizer_kwargs=None,
                  clip_norm=None,
                  min_lr=0,
                  patience=0,
                  factor=0.5,
-                 logging_period=50,
+                 logging_period=100,
                  resume=None,
-                 no_impr=6):
+                 no_impr=8):
         if not th.cuda.is_available():
             raise RuntimeError("CUDA device unavailable...exist")
         if not isinstance(gpuid, tuple):
             gpuid = (gpuid, )
         self.device = th.device("cuda:{}".format(gpuid[0]))
-        self.gpu = gpuid
+        self.gpuid = gpuid
         if checkpoint and not os.path.exists(checkpoint):
             os.makedirs(checkpoint)
         self.checkpoint = checkpoint
@@ -106,7 +108,7 @@ class Trainer(object):
         self.clip_norm = clip_norm
         self.logging_period = logging_period
         self.cur_epoch = 0  # zero based
-        self.no_impr = no_impr
+        self.no_impr = no_impr  # early stop
 
         if resume:
             if not os.path.exists(resume):
@@ -131,13 +133,13 @@ class Trainer(object):
             patience=patience,
             min_lr=min_lr,
             verbose=True)
-        num_params = sum([param.nelement()
-                          for param in nnet.parameters()]) / 10.0**6
+        self.num_params = sum(
+            [param.nelement() for param in nnet.parameters()]) / 10.0**6
 
         # logging
         self.logger.info("Model summary:\n{}".format(nnet))
-        self.logger.info("Loading model({:.2f}M) to GPU:{}".format(
-            num_params, gpuid))
+        self.logger.info("Loading model to GPUs:{}, #param: {:.2f}M".format(
+            gpuid, self.num_params))
         if clip_norm:
             self.logger.info(
                 "Gradient clipping by {}, default L2".format(clip_norm))
@@ -177,9 +179,7 @@ class Trainer(object):
 
     def train(self, data_loader):
         self.nnet.train()
-
-        stats = AverageMeter()
-        timer = SimpleTimer()
+        reporter = ProgressReporter(self.logger, period=self.logging_period)
 
         for egs in data_loader:
             # load to gpu
@@ -187,100 +187,99 @@ class Trainer(object):
 
             self.optimizer.zero_grad()
             loss = self.compute_loss(egs)
-
-            stats.add("loss", loss.item())
             loss.backward()
-
-            progress = stats.count("loss")
-            if not progress % self.logging_period:
-                self.logger.info("Processed {:d} batches...".format(progress))
-
             if self.clip_norm:
                 clip_grad_norm_(self.nnet.parameters(), self.clip_norm)
             self.optimizer.step()
 
-        return stats.value("loss"), stats.count("loss"), timer.elapsed()
+            reporter.add(loss.item())
+        return reporter.report()
 
     def eval(self, data_loader):
+        self.logger.info("Set eval mode...")
         self.nnet.eval()
 
-        stats = AverageMeter()
-        timer = SimpleTimer()
+        reporter = ProgressReporter(self.logger, period=self.logging_period)
 
         with th.no_grad():
             for egs in data_loader:
                 egs = load_obj(egs, self.device)
                 loss = self.compute_loss(egs)
-                stats.add("loss", loss.item())
+                reporter.add(loss.item())
 
-        return stats.value("loss"), stats.count("loss"), timer.elapsed()
+        return reporter.report(details=True)
 
     def run(self, train_loader, dev_loader, num_epochs=50):
-        # using target device
-        with th.cuda.device(self.gpu[0]):
-            stats = dict()
-            no_impr = 0
-            # check if save is OK
+        # avoid alloc memory from gpu0
+        # or use with th.cuda.device(self.gpuid[0]):
+        th.cuda.set_device(self.gpuid[0])
+        memo = dict()
+        # check if save is OK
+        self.save_checkpoint(best=False)
+        cv = self.eval(dev_loader)
+        best_loss = cv["loss"]
+
+        self.logger.info("START FROM EPOCH {:d}, LOSS = {:.4f}".format(
+            self.cur_epoch, best_loss))
+        no_impr = 0
+        # make sure not inf
+        self.scheduler.best = best_loss
+        while self.cur_epoch < num_epochs:
+            cur_lr = self.optimizer.param_groups[0]["lr"]
+            memo["title"] = "Loss(time/N, lr={:.3e}) - Epoch {:2d}:".format(
+                cur_lr, self.cur_epoch + 1)
+
+            tr = self.train(train_loader)
+            memo["tr"] = "train = {:+.4f}({:.2f}m/{:d})".format(
+                tr["loss"], tr["cost"], tr["batches"])
+            cv = self.eval(dev_loader)
+            memo["cv"] = "dev = {:+.4f}({:.2f}m/{:d})".format(
+                cv["loss"], cv["cost"], cv["batches"])
+
+            memo["scheduler"] = ""
+            if cv["loss"] > best_loss:
+                no_impr += 1
+                memo["scheduler"] = "| no impr, best = {:.4f}".format(
+                    self.scheduler.best)
+            else:
+                best_loss = cv["loss"]
+                no_impr = 0
+                self.save_checkpoint(best=True)
+            self.logger.info("{title} {tr} | {cv} {scheduler}".format(**memo))
+            # schedule here
+            self.scheduler.step(cv["loss"])
+            # flush scheduler info
+            sys.stdout.flush()
+            # save checkpoint
+            self.cur_epoch += 1
             self.save_checkpoint(best=False)
-            best_loss, _, _ = self.eval(dev_loader)
-            self.logger.info("START FROM EPOCH {:d}, LOSS = {:.4f}".format(
-                self.cur_epoch, best_loss))
-            while self.cur_epoch < num_epochs:
-                stats[
-                    "title"] = "Loss(time/N, lr={:.3e}) - Epoch {:2d}:".format(
-                        self.optimizer.param_groups[0]["lr"],
-                        self.cur_epoch + 1)
-                tr_loss, tr_batch, tr_cost = self.train(train_loader)
-                stats["tr"] = "train = {:+.4f}({:.2f}m/{:d})".format(
-                    tr_loss, tr_cost / 60, tr_batch)
-                cv_loss, cv_batch, cv_cost = self.eval(dev_loader)
-                stats["cv"] = "dev = {:+.4f}({:.2f}m/{:d})".format(
-                    cv_loss, cv_cost / 60, cv_batch)
-                stats["scheduler"] = ""
-                if cv_loss > best_loss:
-                    stats["scheduler"] = "| no impr, best = {:.4f}".format(
-                        self.scheduler.best)
-                    no_impr += 1
-                else:
-                    best_loss = cv_loss
-                    no_impr = 0
-                    self.save_checkpoint(best=True)
+            if no_impr == self.no_impr:
                 self.logger.info(
-                    "{title} {tr} | {cv} {scheduler}".format(**stats))
-                # schedule here
-                self.scheduler.step(cv_loss)
-                # flush scheduler info
-                sys.stdout.flush()
-                # save checkpoint
-                self.cur_epoch += 1
-                self.save_checkpoint(best=False)
-                if no_impr == self.no_impr:
-                    self.logger.info(
-                        "Stop training cause no impr for {:d} epochs".format(
-                            no_impr))
-                    break
-            self.logger.info(
-                "Training for {} epoches done!".format(num_epochs))
+                    "Stop training cause no impr for {:d} epochs".format(
+                        no_impr))
+                break
+        self.logger.info("Training for {:d}/{:d} epoches done!".format(
+            self.cur_epoch, num_epochs))
 
 
-class UpitTrainer(Trainer):
+class PermutationTrainer(Trainer):
     """
-    Trainer on uPIT loss
+    Trainer on utterance-level PIT loss
     """
 
     def __init__(self, *args, **kwargs):
-        super(UpitTrainer, self).__init__(*args, **kwargs)
+        super(PermutationTrainer, self).__init__(*args, **kwargs)
 
     def compute_loss(self, egs):
         """
         For permutation invariant trainer, egs contains:
-            xlen: length of each utts
+            xlen: length of each utterance
             feats: features that feed networks
-            lx: abs(X)
-            ly: [abs(Y1), abs(Y2) ...]
+            lx: linear spectrogram of mixture
+            ly: linear spectrogram/phase sensitive magnitude of reference speakers
         """
         spk_masks = th.nn.parallel.data_parallel(
-            self.nnet, egs["feats"], device_ids=self.gpu)
+            self.nnet, egs["feats"], device_ids=self.gpuid)
 
         num_utts = egs["xlen"].size(0)
         num_spks = len(egs["ly"])
