+++ egs/upit/nnet/train.py
@@ -7,19 +7,20 @@ import pprint
 import argparse
 import random
 
-from nnet import UpitNet
-from trainer import UpitTrainer
-from utils import dump_json, get_logger
-from dataset import make_pitloader
+from libs.trainer import PermutationTrainer
+from libs.utils import dump_json, get_logger
+from libs.dataset import make_pitloader
+
+from nnet import Nnet
 from conf import trainer_conf, nnet_conf, feats_conf, train_data, dev_data
 
 logger = get_logger(__name__)
 
 
 def run(args):
-    nnet = UpitNet(**nnet_conf)
+    nnet = Nnet(**nnet_conf)
 
-    trainer = UpitTrainer(
+    trainer = PermutationTrainer(
         nnet, gpuid=args.gpu, checkpoint=args.checkpoint, **trainer_conf)
 
     for conf, fname in zip([nnet_conf, feats_conf, trainer_conf],
@@ -31,18 +32,23 @@ def run(args):
         train_data["linear_x"],
         feats_conf,
         train_data,
-        num_utts=args.batch_size)
+        batch_size=args.batch_size,
+        cache_size=args.cache_size)
     feats_conf["shuf"] = False
     dev_loader = make_pitloader(
-        dev_data["linear_x"], feats_conf, dev_data, num_utts=args.batch_size)
+        dev_data["linear_x"],
+        feats_conf,
+        dev_data,
+        batch_size=args.batch_size,
+        cache_size=args.cache_size)
 
     trainer.run(train_loader, dev_loader, num_epochs=args.epochs)
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
-        description="Command to do permutate invariant training, "
-        "auto configured from conf.py",
+        description="Command to do train (B)LSTM with utterance-level "
+        "permutation invariant training, auto configured from conf.py",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument(
         "--gpu", type=int, default=0, help="Training on which GPUs")
@@ -51,13 +57,18 @@ if __name__ == "__main__":
     parser.add_argument(
         "--checkpoint",
         type=str,
-        default="exp/upit/1a",
+        required=True,
         help="Directory to dump models")
     parser.add_argument(
         "--batch-size",
         type=int,
         default=16,
         help="Number of utterances in each batch")
+    parser.add_argument(
+        "--cache-size",
+        type=int,
+        default=8,
+        help="Number of batches cached in the queue")
     args = parser.parse_args()
     logger.info("Arguments in command:\n{}".format(pprint.pformat(vars(args))))
     run(args)
