+++ local/README.md
@@ -1,0 +1,1 @@
+Those scripts work with early version of this repo, which are out of date now.
